Refactor UserEmailAddress.getExperiments
