+++ kifi-backend/modules/shoebox/app/com/keepit/model/UserEmailAddress.scala
@@ -38,23 +38,6 @@ case class UserEmailAddress(
 
 object UserEmailAddress {
   private lazy val random = new SecureRandom()
-  private val kifiDomains = Set("kifi.com", "42go.com")
-  private val testDomains = Set("tfbnw.net", "mailinator.com") // tfbnw.net is for fake facebook accounts
-  private val tagRe = """(?<=\+)[^@+]*(?=(?:\+|$))""".r
-
-  def getExperiments(email: UserEmailAddress): Set[UserExperimentType] = {
-    val Array(local, host) = email.address.address.split('@')
-    val tags = tagRe.findAllIn(local).toSet
-    if (kifiDomains.contains(host) && tags.exists(_.startsWith("autogen"))) {
-      Set(FAKE, AUTO_GEN)
-    } else if (kifiDomains.contains(host) && tags.exists { t => t.startsWith("test") || t.startsWith("utest") }) {
-      Set(FAKE)
-    } else if (testDomains.contains(host)) {
-      Set(FAKE)
-    } else {
-      Set.empty
-    }
-  }
 
   // primary: trueOrNull in db
   def applyFromDbRow(
